MyIoTProject/requirements.txt
Flask==2.3.3
pyserial==3.5
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2